Implements generating font texture atlases
This commit modifies the TextHandler to generate a texture atlas for the
given font. This means that there is one big texture for all characters,
rather than one texture per character, dramatically reducing the number
of textures needed.

These changes also necessitated changes to the handling of vertex
buffers for the text, which also required making the buffer manager's
queued buffer freeing smarter. Now, buffers should be marked with the
index of the last frame that depended on them when they are queued to be
freed, so that they are only freed after that frame finishes.